refator: rename Home folder to HomePage
