+++ client/src/pages/HomePage/carousel/CarouselIndicators.tsx
