+++ client/src/pages/HomePage/carousel/slides/GeneralAssembly.tsx
