+++ client/src/pages/HomePage/carousel/slides/GeneralAssemblyV2.tsx
