+++ client/src/pages/HomePage/carousel/slides/ViewProjects.tsx
