+++ client/src/pages/HomePage/carousel/slides/ViewProjectsV2.tsx
