+++ client/src/pages/HomePage/components/LatestNews.tsx
